Hotel Reservation GUI
Using object-oriented programming techniques and exception handling to implement a Greensboro Hotel Reservation System.  The Greensboro Hotel has 5 rooms, and each room has a number associated with it. The Greensboro Hotel Reservation System must satisfy the following requirements:

A constructor sets the November 26 and 27 and number of rooms.
A method HotelFull to check if there are any rooms available for the customer requested date.
A person reserves a room via ReserveRoom method.
If a customer tries to reserve a room but all rooms are reserved a NoRoom exception will occur. The corresponding exception handler must display related information that must include date and month that the customer requested.
If a customer reserves a room the date, month and room number should be displayed. 
If a customer cancelled the reservation the room will be available.