EBNF Parser
Language implementation systems must input a file, identify substrings, and find all syntax errors, regardless of the specific implementation approach. For an input program a lexical analyzer will break a file into lexemes and tokens. A parser will find all syntax errors. For each error the parser will produce an appropriate diagnostic message and recover quickly.
For the EBNF Described Language defined as follows, choose a programming language that you are familiar with, such as C, C++ or Java, to implement a lexical analyzer and a top-down parser. The lexical analyzer will read the test program and generate lexemes with token types. The parser must use the recursive-descending parsing algorithm to implement the syntax analysis. The project must implement the following functions:
1.	Input the test program.
2.	Use the Lexical Analyzer to generate lexemes and tokens, and display them.
3.	Use the recursive-descending parser to check the syntax of the test program.
4.	If the test program is generated by the EBNF Described Language display “The program is generated by the grammar”, otherwise display “The program cannot be generated by the EBNF Described Language” and identify the syntax error.